# api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Access",
    "version": "Not Defined"
  },
  "paths": {
...
    "/api/audit/users": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Users And Groups",
        "operationId": "users_and_groups_api_audit_users_get",
        "parameters": [
          {
            "name": "q_args",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SearchUserGroupAuditQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TypeVar_Customized_AuditUserGroupRow_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Users",
        "operationId": "users_api_users_get",
        "parameters": [
          {
            "name": "q_args",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SearchUserQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TypeVar_Customized_OktaUserSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "User By Id",
        "operationId": "user_by_id_api_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OktaUserDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}/audit": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "User Audit By Id",
        "operationId": "user_audit_by_id_api_users__user_id__audit_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuditOrderBy": {
        "type": "string",
        "enum": [
          "moniker",
          "created_at",
          "ended_at"
        ],
        "title": "AuditOrderBy"
      },
      "AuditUserGroupRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "group_id": {
            "type": "string",
            "title": "Group Id"
          },
          "role_group_map_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Group Map Id"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner"
          },
          "should_expire": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Should Expire"
          },
          "created_reason": {
            "type": "string",
            "title": "Created Reason",
            "default": ""
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_UserSummaryForAudit"
              },
              {
                "type": "null"
              }
            ]
          },
          "group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_GroupRefForAudit"
              },
              {
                "type": "null"
              }
            ]
          },
          "role_group_mapping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_RoleGroupMappingForAudit"
              },
              {
                "type": "null"
              }
            ]
          },
          "access_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_AccessRequestRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_UserSummaryForAudit"
              },
              {
                "type": "null"
              }
            ]
          },
          "ended_actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_UserSummaryForAudit"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "group_id",
          "is_owner"
        ],
        "title": "AuditUserGroupRow",
        "description": "Mirrors `_serialize_user_group_member` (audit.py:190-226)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OktaUserDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile"
          },
          "manager": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserManagerRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_group_memberships": {
            "items": {
              "$ref": "#/components/schemas/OktaUserGroupMemberDetail"
            },
            "type": "array",
            "title": "Active Group Memberships"
          },
          "active_group_ownerships": {
            "items": {
              "$ref": "#/components/schemas/OktaUserGroupMemberDetail"
            },
            "type": "array",
            "title": "Active Group Ownerships"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "OktaUserDetail"
      },
      "OktaUserGroupMemberDetail": {
        "properties": {
          "is_owner": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Owner"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "created_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Reason",
            "default": ""
          },
          "should_expire": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Should Expire"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OktaUserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_GroupRefForMembership"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_GroupRefForMembership"
              },
              {
                "type": "null"
              }
            ]
          },
          "role_group_mapping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_RoleGroupMappingForMembership"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_role_group_mapping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_RoleGroupMappingForMembership"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "created_at"
        ],
        "title": "OktaUserGroupMemberDetail"
      },
      "OktaUserManagerRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "OktaUserManagerRef",
        "description": "Embedded manager reference inside `OktaUserDetail`.\n\nFlask's `OktaUserSchema.manager = Nested(OktaUserSchema, exclude=(...))`\nretained `profile` (filtered to `USER_DISPLAY_CUSTOM_ATTRIBUTES`). The\nReact user-detail page reads `manager.profile.Title` to render the\nmanager's job title alongside their name."
      },
      "OktaUserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "OktaUserSummary"
      },
...
      "Page_TypeVar_Customized_AuditUserGroupRow_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditUserGroupRow"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[TypeVar]Customized[AuditUserGroupRow]"
      },
...
      "Page_TypeVar_Customized_OktaUserSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OktaUserSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[TypeVar]Customized[OktaUserSummary]"
      },
...
      "SearchUserGroupAuditQuery": {
        "properties": {
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "owner": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "needs_review": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needs Review"
          },
          "order_by": {
            "$ref": "#/components/schemas/AuditOrderBy",
            "default": "created_at"
          },
          "order_desc": {
            "type": "boolean",
            "title": "Order Desc",
            "default": true
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "app_owner": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Owner"
          },
          "managed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "direct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direct"
          },
          "deleted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted"
          }
        },
        "type": "object",
        "title": "SearchUserGroupAuditQuery"
      },
      "SearchUserQuery": {
        "properties": {
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          }
        },
        "type": "object",
        "title": "SearchUserQuery"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "_AccessRequestRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "_AccessRequestRef",
        "description": "Mirrors `_access_request_ref` at audit.py:184-187."
      },
...
      "_GroupRefForAudit": {
        "additionalProperties": true,
        "type": "object"
      },
...
      "_RoleGroupMappingForAudit": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "role_group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_RoleGroupRefForAudit"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "_RoleGroupMappingForAudit",
        "description": "Mirrors `_role_group_mapping_ref` at audit.py:158-165."
      },
      "_RoleGroupMappingForMembership": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "active_role_group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_RoleGroupRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "role_group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_RoleGroupRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "created_at"
        ],
        "title": "_RoleGroupMappingForMembership"
      },
...
      "_UserSummaryForAudit": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "_UserSummaryForAudit",
        "description": "Mirrors `_user_summary` at audit.py:106-117. Subset of\n`OktaUserSummary` from core_schemas.py \u2014 no `updated_at`."
      }
    }
  }
}